+++ internal/system/process_unix.go
@@ -6,6 +6,7 @@
 package system
 
 import (
+	"errors"
 	"os"
 	"os/exec"
 	"syscall"
@@ -19,4 +20,15 @@ func SetProcessGroup(cmd *exec.Cmd) {
 // KillGroup kills the process group
 func KillGroup(process *os.Process) error {
 	return syscall.Kill(-process.Pid, syscall.SIGKILL)
+}
+
+// ProcessExists reports whether a process with the given pid is running.
+// EPERM means the process exists but is owned by another user
+func ProcessExists(pid int) bool {
+	process, err := os.FindProcess(pid)
+	if err != nil {
+		return false
+	}
+	err = process.Signal(syscall.Signal(0))
+	return err == nil || errors.Is(err, syscall.EPERM)
 }